pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Clissman J. Camacho <clissmancamacho@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
fastapi-sqlalchemy = "^0.2.1"
pydantic = "^2.0.3"
alembic = "^1.11.1"
psycopg2 = "^2.9.6"
uvicorn = "^0.23.1"
python-dotenv = "^1.0.0"
asgi-correlation-id = "^4.2.0"
flake8 = "^6.0.0"
black = "^23.7.0"
loguru = "^0.7.0"
psycopg2-binary = "^2.9.6"
asyncpg = "^0.28.0"
greenlet = "^2.0.2"
python-multipart = "^0.0.6"
httpx = "^0.24.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
coverage = "^7.2.7"
pylist = "^1.4.0"
pytest-cov = "^4.1.0"
asgi-lifespan = "^2.1.0"
pytest-asyncio = "^0.21.1"
debugpy = "^1.6.7"
autopep8 = "^2.0.2"
isort = "^5.12.0"

[tool.isort]
profile = "black"
src_paths = ["app", "tests"]
combine_as_imports = true

[tool.black]
line-length = 200
target-version = ["py310"]
include = '\.pyi?$'

[tool.flake8]
ignore = ["E266"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
